refact: rename ghostos.__init__ to ghostos.bootstrap, remove streamlit testing design1
