--- ghostos/prototypes/streamlitapp/design1/README.md
@@ -1,4 +1,0 @@
-# About this directory
-
-The raw design of the ghostos streamlit application. 
-remove later.
