--- ghostos/prototypes/streamlitapp/design1/resources/knowledge.py
@@ -1,6 +1,0 @@
-import streamlit as st
-
-st.write(st.query_params.to_dict())
-
-with st.chat_message("assistant"):
-    st.write("hello world!")
